refactor(components): moved EntityInfos out of widgets into components
